+++ dotbot_firefox.py
@@ -1,5 +1,5 @@
 # dotbot-firefox -- Configure your Firefox profile(s) using dotbot.
-# Copyright 2022 Kurt McKee <contactme@kurtmckee.org>
+# Copyright 2022-2023 Kurt McKee <contactme@kurtmckee.org>
 # SPDX-License-Identifier: MIT
 
 
@@ -14,7 +14,7 @@ import typing
 import dotbot.plugin
 import dotbot.plugins.link
 
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 
 VALID_DIRECTIVES: set[str] = {"firefox"}
 
@@ -46,6 +46,17 @@ def _get_profile_directories() -> typing.Iterable[pathlib.Path]:
         snap_user_common = "~/snap/firefox/common"
         defaults.append(f"{snap_user_common}/.mozilla/firefox")
 
+        # When Firefox is installed and run as a flatpak,
+        # it stores its profile info under this directory:
+        #
+        #     $HOME/.var/app/org.mozilla.firefox
+        #
+        # $HOME can be customized inside the flatpak sandbox environment,
+        # but it seems unlikely that this level of customization will be needed.
+        #
+        flatpak_home = "~/.var/app/org.mozilla.firefox"
+        defaults.append(f"{flatpak_home}/.mozilla/firefox")
+
     for default in defaults:
         path = pathlib.Path(os.path.expandvars(os.path.expanduser(default)))
         if not path.is_dir():
@@ -56,7 +67,13 @@ def _get_profile_directories() -> typing.Iterable[pathlib.Path]:
                 yield profile
 
 
-class Firefox(dotbot.plugin.Plugin):
+# mypy 1.3.0 reports the following error for the Firefox class:
+#
+#   Class cannot subclass "Plugin" (has type "Any")  [misc]
+#
+# The "type: ignore[misc]" comment below suppresses this specific error.
+#
+class Firefox(dotbot.plugin.Plugin):  # type: ignore[misc]
     def can_handle(self, directive: str) -> bool:
         """
         Flag whether this plugin supports the given *directive*.
@@ -95,4 +112,4 @@ class Firefox(dotbot.plugin.Plugin):
             log.warning("No Firefox profiles found")
             return True
 
-        return link_plugin.handle("link", links)
+        return bool(link_plugin.handle("link", links))