+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
 ..  dotbot-firefox -- Configure your Firefox profile(s) using dotbot.
-..  Copyright 2022 Kurt McKee <contactme@kurtmckee.org>
+..  Copyright 2022-2023 Kurt McKee <contactme@kurtmckee.org>
 ..  SPDX-License-Identifier: MIT
 
 
@@ -14,6 +14,21 @@ Unreleased changes to the code are documented in
 in the ``changelog.d/`` directory on GitHub.
 
 ..  scriv-insert-here
+
+.. _changelog-1.1.0:
+
+1.1.0 — 2023-05-23
+==================
+
+Python support
+--------------
+
+-   Add support for Python 3.12.
+
+Added
+-----
+
+-   Support Firefox Flatpak for Linux profile directories.
 
 .. _changelog-1.0.1: